GOV-31148 Rewrite query to improve performance Include Create Indexes, Rewrite Query, Add condition.

+++ d360.model/helpers/filters/FilterExpressionParser.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
@@ -631,23 +632,31 @@ namespace d360.model.helpers
 				}
 			}
 
-			foreach (var scoreField in fieldTypes.Where(x => x.Type == DataType.Score.ToString()))
+			foreach (var token in filterTokens)
 			{
-				string aggregatedDataSQL = @$"
-				drop table if exists #scoreTempValues{scoreField.ID}
-				create table #scoreTempValues{scoreField.ID} (AssetId int, FormattedValue decimal(8,3))
-				create index ix_scoreTempValues{scoreField.ID} on #scoreTempValues{scoreField.ID}	(AssetId);
+				if (token is TempTableFieldToken)
+				{
+					var tokenConvert = (TempTableFieldToken)token;
+					if (tokenConvert.CurrentFieldType == DataType.Score.ToString().ToLower(CultureInfo.InvariantCulture))
+					{
+						var scoreField = fieldTypes.Where(x => x.Name.ToLower(CultureInfo.InvariantCulture) == tokenConvert.Field.ToLower(CultureInfo.InvariantCulture)).FirstOrDefault();
+						string aggregatedDataSQL = @$"
+								drop table if exists #scoreTempValues{scoreField.ID}
+								create table #scoreTempValues{scoreField.ID} (AssetId int, FormattedValue decimal(8,3))
+								create index ix_scoreTempValues{scoreField.ID} on #scoreTempValues{scoreField.ID}	(AssetId);
 
-				insert into #scoreTempValues{scoreField.ID}(AssetId, FormattedValue)
-				select a.id, S.Value*100 from Asset A
-				inner join metrics.Score S on S.AssetUid = A.uid and S.EndDate is null 
-				inner join metrics.Allocation Al on Al.Uid = S.AllocationUid and Al.ScoreType = {(int)scoreField.ScoreType} and Al.OverrideName is null
-				where A.AssetTypeID = @assettypeid";
+								insert into #scoreTempValues{scoreField.ID}(AssetId, FormattedValue)
+								select a.id, S.Value*100 from Asset A
+								inner join metrics.Score S on S.AssetUid = A.uid and S.EndDate is null 
+								inner join metrics.Allocation Al on Al.Uid = S.AllocationUid and Al.ScoreType = {(int)scoreField.ScoreType} and Al.OverrideName is null
+								where A.AssetTypeID = @assettypeid";
 
-				data.Add(new AdvancedFilterTempTableInfo { TempTableQuery = aggregatedDataSQL });
+						data.Add(new AdvancedFilterTempTableInfo { TempTableQuery = aggregatedDataSQL });
+					}
+				}
 			}
 
-			foreach(var token in filterTokens)
+			foreach (var token in filterTokens)
 			{
 				if (token is ITempTableFilter)
 				{
